refactor: Kafka 이벤트 레코드를 application 계층으로 이동 - KafkaTopics 내 이벤트 레코드 5개를 application/event/kafka/ 패키지로 분리 - KafkaTopics는 토픽 상수만 유지 - ProductCommandService import를 application.event.kafka.* 로 변경
