--- src/main/java/com/trustamarket/productservice/application/event/InspectionRejectedEvent.java
@@ -1,12 +1,0 @@
-package com.trustamarket.productservice.application.event;
-
-import com.trustamarket.productservice.domain.product.Product;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class InspectionRejectedEvent {
-    private final Product product;
-    private final String reason;
-}
