--- src/main/java/com/trustamarket/productservice/application/event/ProductDeletedEvent.java
@@ -1,12 +1,0 @@
-package com.trustamarket.productservice.application.event;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-import java.util.UUID;
-
-@Getter
-@RequiredArgsConstructor
-public class ProductDeletedEvent {
-    private final UUID productId;
-}
