+++ src/main/java/com/trustamarket/productservice/application/event/kafka/InspectionPriceAcceptedEvent.java
@@ -1,0 +1,10 @@
+package com.trustamarket.productservice.application.event.kafka;
+
+import java.util.UUID;
+
+public record InspectionPriceAcceptedEvent(
+        UUID eventId,
+        UUID productId,
+        UUID sellerId,
+        Long finalPrice
+) {}
