+++ src/main/java/com/trustamarket/productservice/application/event/kafka/InspectionPriceRejectedEvent.java
@@ -1,0 +1,10 @@
+package com.trustamarket.productservice.application.event.kafka;
+
+import java.util.UUID;
+
+public record InspectionPriceRejectedEvent(
+        UUID eventId,
+        UUID productId,
+        UUID sellerId,
+        String reason
+) {}
