+++ src/main/java/com/trustamarket/productservice/application/event/kafka/InspectionRequestedEvent.java
@@ -1,0 +1,12 @@
+package com.trustamarket.productservice.application.event.kafka;
+
+import java.util.UUID;
+
+public record InspectionRequestedEvent(
+        UUID eventId,
+        UUID productId,
+        UUID sellerId,
+        UUID centerId,
+        long originalPriceAmount,
+        String currency
+) {}
