+++ src/main/java/com/trustamarket/productservice/application/event/kafka/ProductCreatedEvent.java
@@ -1,0 +1,12 @@
+package com.trustamarket.productservice.application.event.kafka;
+
+
+import java.util.UUID;
+
+public record ProductCreatedEvent(
+        UUID productId,
+        UUID sellerId,
+        UUID categoryId,
+        Long price,
+        String inspectionStatus
+) {}
