+++ src/main/java/com/trustamarket/productservice/application/event/kafka/ProductDeletedEvent.java
@@ -1,0 +1,7 @@
+package com.trustamarket.productservice.application.event.kafka;
+
+import java.util.UUID;
+
+public record ProductDeletedEvent(
+        UUID productId
+) {}
