+++ src/main/java/com/trustamarket/productservice/infrastructure/kafka/KafkaTopics.java
@@ -10,38 +10,4 @@ public class KafkaTopics {
     public static final String INSPECTION_PRICE_ACCEPTED_TOPIC = "inspection.price.accepted";
     public static final String INSPECTION_PRICE_REJECTED_TOPIC = "inspection.price.rejected";
 
-    public record ProductCreatedEvent(
-            UUID productId,
-            UUID sellerId,
-            UUID categoryId,
-            Long price,
-            String inspectionStatus
-    ) {}
-
-    public record ProductDeletedEvent(
-            UUID productId
-    ) {}
-
-    public record InspectionRequestedEvent(
-            UUID eventId,
-            UUID productId,
-            UUID sellerId,
-            UUID centerId,
-            long originalPriceAmount,
-            String currency
-    ) {}
-
-    public record InspectionPriceAcceptedEvent(
-            UUID eventId,
-            UUID productId,
-            UUID sellerId,
-            Long finalPrice
-    ) {}
-
-    public record InspectionPriceRejectedEvent(
-            UUID eventId,
-            UUID productId,
-            UUID sellerId,
-            String reason
-    ) {}
 }
